requirements.txt
<<<<<<< HEAD
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Mail
python-dotenv
email-validator
itsdangerous
python-dateutil
pytest
selenium
webdriver-manager
=======
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-Migrate
python-dotenv
email-validator
itsdangerous
python-dateutil
>>>>>>> main
